IHS-156 refactor the test cases to have a better view on the tests perimeter
